:tada: security done

+++ crud/commandLines/delete.go
@@ -14,27 +14,32 @@ func Delete(w http.ResponseWriter, r *http.Request) {
 	utils.SetCorsHeaders(&w)
 
 	if r.Method == "DELETE" {
-		//connect the database
-		db := utils.DbConnect()
 
-		//get url values
-		vars := mux.Vars(r)
-		id_commandLine := vars["id_commandLine"]
+		if utils.Securized(w, r) {
+			//connect the database
+			db := utils.DbConnect()
 
-		//create the sql query
-		sqlQuery := ("DELETE FROM command_line WHERE id = " + id_commandLine + ";")
+			//get url values
+			vars := mux.Vars(r)
+			id_commandLine := vars["id_commandLine"]
 
-		//execute the sql query
-		_, err := db.Exec(sqlQuery)
-		utils.CheckErr(err)
+			//create the sql query
+			sqlQuery := ("DELETE FROM command_line WHERE id = " + id_commandLine + ";")
 
-		//close the database connection
-		db.Close()
+			//execute the sql query
+			_, err := db.Exec(sqlQuery)
+			utils.CheckErr(err)
 
-		//create the json response
-		json.NewEncoder(w).Encode(structures.JsonReponseCommandLines{
-			Type:    "success",
-			Message: "Command line deleted",
-		})
+			//close the database connection
+			db.Close()
+
+			//create the json response
+			json.NewEncoder(w).Encode(structures.JsonReponseCommandLines{
+				Type:    "success",
+				Message: "Command line deleted",
+			})
+		} else {
+			w.WriteHeader(http.StatusUnauthorized)
+		}
 	}
 }
